backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.25
